feat(usb): improve ADBUSBDevice layering
Preparing work for webusb integration.
- ADBUSBDevice does not depend anymore on rusb code, but now holds an
USBTransport implementor as a TRANSPORT generic.
- This trait wraps ADBTransport + ADBMessageTransport to "hide" concrete
dependencies and is currently implemented by WiredUSBTransport.
