+++ adb_client/Cargo.toml
@@ -21,6 +21,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 default = []
 mdns = ["dep:mdns-sd"]
 usb = ["dep:rusb"]
+# webusb = ["dep:webusb-web"]
 
 [dependencies]
 base64 = { version = "0.22.1" }
@@ -53,6 +54,9 @@ mdns-sd = { version = "0.19.1", default-features = false, features = [
 # USB-only dependencies
 rusb = { version = "0.9.4", features = ["vendored"], optional = true }
 num_enum = { version = "0.7.6" }
+#########
+# `webusb` feature-specific dependencies
+# webusb-web = { version = "0.5.1", optional = true }
 #########
 
 [dev-dependencies]
