+++ adb_client/src/message_devices/usb/adb_device_info.rs
@@ -1,0 +1,10 @@
+/// Represents an Android device connected via USB
+#[derive(Clone, Debug)]
+pub struct ADBDeviceInfo {
+    /// Vendor ID of the device
+    pub vendor_id: u16,
+    /// Product ID of the device
+    pub product_id: u16,
+    /// Textual description of the device
+    pub device_description: String,
+}
