+++ adb_client/src/message_devices/usb/mod.rs
@@ -1,7 +1,11 @@
+#![doc = include_str!("./README.md")]
+
+mod adb_device_info;
 mod adb_usb_device;
 mod usb_transport;
-mod utils;
+mod wired_usb_transport;
 
+pub use adb_device_info::ADBDeviceInfo;
 pub use adb_usb_device::ADBUSBDevice;
 pub use usb_transport::USBTransport;
-pub use utils::{ADBDeviceInfo, find_all_connected_adb_devices};
+pub use wired_usb_transport::WiredUSBTransport;
